fix: add uint8_t cast on sizeof(buf) in nRF24 payload length guard
Agent-Logs-Url: https://github.com/maxim75/cc1101_receiver_esp32/sessions/2e9f2deb-e697-4acf-9f65-a31d00ca5fc1

+++ src/main.cpp
@@ -307,7 +307,7 @@ void loop() {
         uint8_t buf[32] = {};
         uint8_t len = nrf24.getDynamicPayloadSize();
         if (len == 0 || len > sizeof(buf)) len = nrf24.getPayloadSize();
-        if (len == 0 || len > sizeof(buf)) len = sizeof(buf);
+        if (len == 0 || len > sizeof(buf)) len = (uint8_t)sizeof(buf);
         nrf24.read(buf, len);
 
         fillPacket(lastPacket, "nRF24", buf, len, 0.0f, false);
